feat:系统管理-菜单管理-编辑、删除

+++ model/src/main/java/com/autumn/blog/model/entity/system/Menu.java
@@ -44,5 +44,5 @@ public class Menu extends BaseEntity {
     // 外链地址
     String redirect;
     // 菜单类型
-    Long menuType;
+    Integer menuType;
 }
